ZERO-125 [refactor] : 컴포너는 분리
