+++ src/components/@shared/NavBar.tsx
@@ -1,32 +1,15 @@
 import Dropdown, { Option } from '@components/@shared/Dropdown';
-import Image from 'next/image';
-import PCLogo from 'public/images/logo_pc.png';
 import UserIcon from 'public/icons/user.svg';
-import ArrowDown from 'public/icons/arrow_down.svg';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
-import Link from 'next/link';
-import Menu from 'public/icons/menu.svg';
-import { useModal } from '@hooks/useModal';
-import AddTeamModal from '@components/team/AddTeamModal';
-import KebabIcon from 'public/icons/kebab_small.svg';
 import { useUserData } from '@hooks/mysetting/useUserData';
-import Button from './Button';
+import NavBarTeam from './NavBar_Team';
 
 export default function NavBar() {
   const router = useRouter();
   const [isLogoOnlyPage, setIsLogoOnlyPage] = useState(false);
-  const [selectedTeam, setSelectedTeam] = useState<Option | null>(null);
   const [isClient, setIsClient] = useState(false);
-
   const { data } = useUserData();
-
-  const {
-    isOpen: addIsOpen,
-    onOpen: addOpenModal,
-    onClose: addCloseModal,
-  } = useModal();
-
   const handleLogout = () => {
     localStorage.removeItem('userStorage');
     router.push('/signin');
@@ -84,93 +67,16 @@ export default function NavBar() {
     },
   ];
 
-  const handleTeamSelet = (groupId: number) => {
-    router.push(`/${groupId}`);
-  };
-
-  const teams: Option[] = [
-    ...(data?.memberships?.map((membership) => ({
-      label: membership.group.name || '',
-      component: (
-        <div className="flex items-center justify-between">
-          <div
-            className="flex items-center gap-3"
-            onClick={() => handleTeamSelet(membership.groupId)}
-          >
-            <Image
-              src={membership.group.image}
-              alt="팀 이미지"
-              width={30}
-              height={30}
-              className="rounded-md"
-            />
-            {membership.group.name}
-          </div>
-          <KebabIcon />
-        </div>
-      ),
-    })) || []),
-    {
-      label: '팀 메뉴',
-      component: (
-        <Button bgColor="transparent" border="white" size="full" height={40}>
-          + 팀 추가하기
-        </Button>
-      ),
-    },
-  ];
-
-  const handleSelectTeam = (option: Option) => {
-    if (option.label === '팀 메뉴') {
-      addOpenModal();
-    } else {
-      setSelectedTeam(option);
-    }
-  };
+  if (!data) {
+    // data가 없을 경우 컴포넌트 렌더링을 하지 않음
+    return null;
+  }
 
   return (
     <header className=" flex h-16 items-center justify-center border-b border-border-primary border-opacity-10 bg-background-secondary px-6">
       <nav className="flex h-8 w-[1200px]  items-center justify-between text-text-primary max-xl:w-full max-md:w-full ">
         <div className="flex items-center gap-10 max-md:gap-5">
-          <div className="md:hidden">
-            <Dropdown
-              options={teams}
-              selected={selectedTeam}
-              onSelect={handleSelectTeam}
-              triggerClass="flex gap-[12px] items-center text-text-primary"
-              triggerIcon={<Menu />}
-              optionsWrapClass="mt-[30px] flex p-[16px] rounded-[12px]"
-              optionClass="px[8px] py-[7px] rounded-[8px] w-[186px] h-[46px] hover:bg-background-tertiary"
-            />
-          </div>
-          <Link href="/">
-            <div className="block max-xl:hidden">
-              <Image src={PCLogo} alt="로고" width={158} height={32} />
-            </div>
-            <div className="hidden max-xl:block">
-              <Image src={PCLogo} alt="로고" width={102} height={20} />
-            </div>
-          </Link>
-          {!isLogoOnlyPage && (
-            <>
-              <div className="max-md:hidden">
-                <Dropdown
-                  initialOption={teams[0]}
-                  options={teams}
-                  selected={selectedTeam}
-                  onSelect={handleSelectTeam}
-                  triggerClass="flex gap-[12px] items-center text-text-primary"
-                  triggerIcon={<ArrowDown />}
-                  optionsWrapClass="mt-[30px] flex p-[16px] rounded-[12px]"
-                  optionClass="px[8px] py-[7px] rounded-[8px] w-[186px] h-[46px] hover:bg-background-tertiary"
-                />
-                <AddTeamModal isOpen={addIsOpen} closeModal={addCloseModal} />
-              </div>
-              <Link href="/article">
-                <span className="max-md:hidden">자유게시판</span>
-              </Link>
-            </>
-          )}
+          <NavBarTeam data={data} />
         </div>
         {!isLogoOnlyPage && (
           <div>
